Make some fields unique. Make the catalog route and the books route use book IDs instead of ISBN10. Make ISBN10 optional.

+++ src/routes/user_routes.rs
@@ -47,7 +47,7 @@ async fn list_users(
 /*
     GET /users/{user_id} - Retorna informações de usuário com base no ID
 */
-async fn get_user_id(Path(user_id): Path<usize>, State(pool): State<DbPool>) -> Json<User> {
+async fn get_user_id(Path(user_id): Path<i64>, State(pool): State<DbPool>) -> Json<User> {
     let conn = pool.get().await.unwrap();
     let row = conn
         .query_one("SELECT * FROM users WHERE id = $1", &[&(user_id as i64)])
